pyproject.toml
[project]
name = "proteinsequencechatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.56",
    "langchain-groq>=0.3.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "tavily-python>=0.7.2",
    "wikipedia>=1.4.0",
]
